Add files via upload
Here is my Personal Training Tracker I had started to make. Written using C++, its original intent was just to familiarize me more with this language as it was still newer to me. Once I got started I switched gears however and began making this record keeping tool that would have been helpful for my current job at the moment. This tool was to be used to keep track of client sessions and help manage the whole process that was currently being only kept track of on paper which always had errors. I was almost done with it, at this point I was just running through it to work out any bugs before I went back to try and make some of the functions more efficient. Sadly the gym I was working at decided to go a different route and became no longer interested in tracking the sessions electronically. As for now, this tracker is staying left in its current state while I focus more on school and my next steps career wise.